Oracle-JavaJRE-8_uninstall
Deployable bash script to remove Java JRE from Macs.

The script looks for the existence of /Library/Internet Plug-Ins/JavaAppletPlugin.plugin/Contents/Enabled.plist to confirm if Java is installed. If the .plist is found the script will identify the version of Java installed and echo that version. If Java is not found the script will report Java is not installed and exit. If Java is found the script will remove the directories recommended for uninstallation by Oracle. 

Once the uninstall completes the script will check for the install of Java once more. If Java is found the script will exit with an error, if Java is not found the script will exit with a success.