Background Services Report
This script will generate a log file on the current users desktop listing the background services. The information logged can be used to create configuration profiles to approve Background Services (Login items) with MDM.

Script is designed to be uploaded in to JAMF Pro, and uses JAMF helper to determine if the user wants to open the log file or not. 

JAMF Helper functions can be removed without harming the function of the script if JAMF is not being used.