Create Application_Last_Used_2302
Initial post.

Function of the script is to be used as an Extension Attribute for JAMF. The script will identify the last time that the application specified in line 3 was opened. If the application has never been opened it will report (null) which will cause most of the function of the script to fail as they are looking for a numeric value. Added 12-16 to attempt to account for the (null) response. The script functions but needs some debugging still.

If response is (null) - the script will echo "Uninstall"
If the application has not be opened in 30 days - the script will echo "Uninstall"
If the application has been opened within the last 30 days - the script will echo "Compliant"

Make a smart group targeting the Extension Attribute with a Is "Uninstall". then make a Policy to uninstall the App on a recurring checkin targeting that smart group.