Create Provision_FileVault_Access_2303
How to use the script:
Replace Line 21's LocalAdminUserNameHere with the the username of an Account with a FileVault Token. The account does not need to have administrative access to macOS.
Replace Line 22’s LocalAdminPasswordHere with the password for the account from line 21.

Script can be uploaded to JAMF, and use JAMF Parameters for the user name and password for line 21 and 22. I recommend you look in to salting the password of the local user account so it is not passed in plane text. Script can also be saved locally and run with terminal. 

MacOS handles FileVault on the user level. To manipulate FileVault you need the user name and password of a FileVault token holder. To give a FileVault token to a user, the user must provide their password. Due to these requirements the script must have the user name and password for a FileVault token holder, and the username and password of the person you want to give the FileVault token to.

Apple’s direction is to eventually retire fdesetup, so this script relies on sysadminctl to pass the FileVault token.