Create Map_Network_Drive
This script was designed to mount a network share on an event such as user login, or through a
self service policy.  Using Jamf's policy engine, a policy can be scoped so that
users and groups (local or directory-based) will mount run this script, and therefore mount the
assigned network share.  For directory-based users, it is recommended to use the "kerberos"
authentication type.  For local users, the "password" authentication type must be used.
#
For kerberos authentication to work properly, the user must be able to manually mount the share
when logged in by navigating to "Go" > "Connect to Server..." in the Finder and the user must be
able to mount the share without authenticating.  To ensure that your directory users are
obtaining kerberos tickets properly, navigate to "System" > "Library" > "CoreServices" and open
the "Kerberos Ticket Viewer" application while a directory user is logged in.